perf: defer studio demo visuals until visible
